Moving the full argocd schema from the client app to myks
