+++ internal/myks/templates/argocd/application.ytt.yaml
@@ -11,7 +11,7 @@
 apiVersion: argoproj.io/v1alpha1
 kind: Application
 metadata:
-  name: #@ app_name
+  name: #@ env_name + "-" + app_name
   namespace: #@ a.namespace
 spec:
   project: #@ project_name
